Consolidate API response types and transformers in shared package
- Add shared/api/responses.ts with API response type definitions
- Add shared/api/transformers.ts with toProject, toRepository, toTask, etc.
- Update shared package.json with exports for types and api
- Refactor frontend to use shared API types and transformers
- Remove duplicated type definitions and transformation functions from frontend
