+++ packages/frontend/src/api/projects.ts
@@ -1,31 +1,6 @@
+import { type ProjectResponse, toProject } from "shared/api";
 import type { Project } from "shared/types";
 import { apiPost } from "./client";
-
-export interface ProjectResponse {
-  id: string;
-  name: string;
-  description: string | null;
-  createdAt: string;
-  updatedAt: string;
-}
-
-export interface RepositoryResponse {
-  id: string;
-  name: string;
-  path: string;
-  defaultBranch: string;
-  createdAt: string;
-  updatedAt: string;
-}
-
-function toProject(response: ProjectResponse): Project {
-  return {
-    ...response,
-    description: response.description ?? undefined,
-    createdAt: new Date(response.createdAt),
-    updatedAt: new Date(response.updatedAt),
-  };
-}
 
 export async function createProject(
   name: string,
